# Close #139 Simplify/optimize the code for current deposition on GPU