--- a/fbpic/openpmd_diag/boosted_particle_diag.py
+++ b/fbpic/openpmd_diag/boosted_particle_diag.py
@@ -667,6 +667,7 @@ class ParticleCatcher:
             Nz, Nr = species.grid_shape
             # Calculate cell area to get particles from
             # - Get z indices of the slices in which to get the particles
+            # (mirrors the index calculation in `get_cell_idx_per_particle`)
             iz_curr = math.ceil((current_z_boost-zmin-0.5*dz)/dz)
             iz_prev = math.ceil((previous_z_boost-zmin-0.5*dz + dt*c)/dz) + 1
             # - Get the prefix sum values that correspond to these indices