This pull request is a proposition to slightly change the way in which we sort the particles on the GPU, in order to simplify the deposition kernels.

More precisely, in the current `dev` branch, the particles are sorted according to the index of their lower grid point in z and r as shown here with tuples of `(iz_lower, ir_lower)`. (In the code, these tuples are actually translated into a single index `ir_lower + Nr*iz_lower`, but this is a detail.)
<img width="400" alt="screen shot 2018-01-09 at 10 27 35 pm" src="https://user-images.githubusercontent.com/6685781/34758635-628d8952-f58c-11e7-8e0a-fd1a4bcad683.png">
Notice that the particles that are below the first radial grid cell do not have a lower grid point in r, and so, **currently**, they are put in the same "bag" as the particles that are inbetween the first radial grid point and second radial grid point. (The same problem does not occur in `z` because of the periodicity in z.)

Then, the deposition kernel on GPU go through the particles in each colored area and deposits charge/current to the surrounding grid points. However, because the particles below and above the first radial grid cell do not deposit charge to the same grid points and in the same way, the deposition kernel need a lot of if conditions to treat this special case.

Instead, I suggest that we slightly change the index according to which the particles are sorted and use, instead, the upper index:
<img width="375" alt="screen shot 2018-01-09 at 10 37 03 pm" src="https://user-images.githubusercontent.com/6685781/34758869-b1af326e-f58d-11e7-8fee-0991739232df.png">
The particles that are above the last grid point do not have an upper grid point, but in that case we can simply use `ir_upper = Nr`. Thus the index `ir_upper` can have values between `0` and `Nr` (included) and so the linearized index should be calculated according to `ir_upper + (Nr+1)*iz_upper`. (The same could have been done with `ir_lower` having values between `-1` and `Nr-1` but is less intuitive because of potentially negative values.)

In this case, the deposition kernels can be significantly simplified because we now know, from the sorting index (a.k.a prefix sum index) to which cells the macroparticles will deposit, without ambiguity. This is what this pull request does.

The automated tests pass on GPU.

The changes show essentially no speed-up (see table below), but allow, on the whole, to remove ~400 lines of code. In addition, I think the kernels are also conceptually slightly easier to understand.

|     | Cubic J  | Cubic rho | Linear J | Linear rho |
|-----|----------|-----------|----------|------------|
| dev | 89.438ms | 22.496ms  | 16.943ms | 5.7967ms   |
| PR  | 89.117ms | 22.413ms  | 16.832ms | 5.7674ms   |

This PR has a lot of changes in the file `cuda_methods.py`. I suggest looking at the kernel for linear deposition of rho to get the idea of the PR (since it is the simplest kernel and the other kernels are just repetitions of the same idea.)