@MKirchen 2018-01-13
Nice! This really simplifies the code a lot!